refactor(browser-repl): move preprocessor in subfolder
