+++ packages/browser-repl/src/lib/interpreter/preprocessor/collect-top-level-lexical-context.spec.ts
