+++ packages/browser-repl/src/lib/interpreter/preprocessor/save-and-restore-lexical-context.ts
