+++ packages/browser-repl/src/lib/interpreter/preprocessor/wrap-in-async-function-call.ts
