+++ packages/browser-repl/src/lib/interpreter/preprocessor/wrap-object-literal.ts
